requirements.txt
requests
python-dotenv
azure-storage-blob
streamlit
pandas
deltalake